Add user-friendly version check to bootstrap file
Most of the plugin uses syntax from PHP 7.1, so this is the minimum version. Provide a version check, admin notice and self-deactivation if the site is running an earlier version of PHP.

The syntax in the main plugin file should be compatible with PHP 5.2.4, so that on attempted activation / update, it doesn't give a White Screen if the site is running on less than the required PHP version. The rest of the plugin initialisation is moved to `init.php`.