Run plugin initialization on `plugins_loaded` hook
When the PHP requirement is not met the plugin begins the deactivation process, first checking if a notice should be displayed to the user.  Apparently it is necessary to hook into the WP event cycle so that the function `wp_get_current_user()` is available. I tested the `plugins_loaded` and `init` hooks. Both worked but I opted to use the earlier hook.

Fixes #7