Check and deactivate correctly if PHP req. not met
While `current_user_can()` is available at the time the plugins are being loaded, one of the later functions, `wp_get_current_user()` is not.

This solution still checks for the PHP level at plugin parsing time, but if it's not sufficient, then functions are defined and hooked in to `plugins_loaded` to deactivate and show a notice.

Fixes #7